add test (print message)
