+++ datadriven/src/line_sparser.rs
@@ -72,9 +72,16 @@ fn split_directives(line: &str) -> Result<Vec<String>> {
 
 #[cfg(test)]
 mod tests {
-    use crate::default_logger;
     use crate::line_sparser::{parse_line, split_directives};
     use anyhow::Result;
+    use slog::Drain;
+
+    fn default_logger() -> slog::Logger {
+        let decorator = slog_term::TermDecorator::new().build();
+        let drain = slog_term::FullFormat::new(decorator).build().fuse();
+        let drain = slog_async::Async::new(drain).build().fuse();
+        slog::Logger::root(drain, o!())
+    }
 
     #[test]
     fn test_parse_line() -> Result<()> {
