+++ datadriven/src/testdata/log_message/data_001.txt
@@ -1,0 +1,17 @@
+debug cmd=example
+----
+DEBUG cmd="example"
+
+info cmd=example
+----
+Please use debug instead.
+
+warn cmd=example
+----
+Please use debug instead.
+Please use debug instead.
+
+error cmd=example
+----
+
+
