+++ datadriven/src/testdata/unknown_data/data_002.txt
@@ -1,5 +1,5 @@
 # data with error
-fibonacci a=3 b=?4
+fibonacci a=3 b?4
 ----
 a=3
 b=5
